Simplifies RabbitMQ credential management
Refactors the RabbitMQ distributed lock service to use a simpler OAuth 2.0 token management approach.

The previous implementation used a `CredentialsRefresher` that proactively refreshed the token in the background. This has been replaced with a `CredentialsProvider` that is called by the RabbitMQ client on each connection attempt. This ensures that a valid token is always used, and simplifies the code by removing the need for a background task.

Also retrieves initial token to pre-populate cache.